Another improvement in wolf AI
The fix I was working on in the last commit didn't really work as expected. I know implemented this simple failsafe to when the wolf gets stuck in walls. I don't consider this to be the final fix I'll be doing in this AI, but it's working pretty solid compared to last fixes