Wolf attack AI completely rewritten
Cleaned the code a lot, it's way more understandable now. Also added the functionality that the wolf stops before jumping, like I said I would add.

TIL that you can have a variable set as undefined, how the heck did I never knew this value existed till now?

The next thing will be a failsafe, to move the wolf towards the player if it gets stuck in a wall, which can happen.

Will be commiting this to main now