Wolf wandering AI and Damage
Gave the wolf the ability to choose if it wants to go to a random place on the map, of course it's not limited to any area yet, but I will add that later, so you don't see wolfs in places they shouldn't be lol

And also added damage, if you hit the wolf 3 times, it will die.

Forgot to code the failsafe today, maybe tomorrow I'll get to it if I'm not tired, although I know how the code will look like and am certain it will work (because I have tested it before)

And also put my debug function back in the game, because for some reason it dissapeared? I don't know what happened, but I had it saved as a package, so no problem.

That's it for today, will probably be unavailable in this weekend, have some relatives I need to visit.
So I'll try to get some stuff done before friday