Map Grid System
Implemented a grid system to deal with maps. I find this type of map easier to deal with, both on the level design and memory effciency sides. And also because I like this system in games.

Probably has a few things to make better, but I'm happy with the results. I would like to thank my math teacher from last year, for teaching me about matrices and making this possible lol

Made some other small changes, as changing the camera size, otherwise it would be too big for the grid. I'll maybe change these values again, since the value I'm using makes the the camera a little stretched because of proportions